+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Unreleased
+
+## Added
+
+## Fixed
+
+## Changed
+
 # 1.3.51 (2022-10-18 / 2d9d7f4)
 
 ## Added