+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 1.4.58 (2023-01-16 / 1446eef)
 
 ## Changed